+++ src/brussels/__tests__/types/file/test_file.py
@@ -5,7 +5,6 @@ from typing import TYPE_CHECKING, cast
 from uuid import UUID
 
 import pytest
-from sqlalchemy import Engine, create_engine
 from sqlalchemy.orm import Mapped, Session, mapped_column
 
 from brussels.base import Base, DataclassBase
@@ -20,8 +19,11 @@ except ImportError:
 
 if TYPE_CHECKING:
     from obstore import Attributes, GetOptions, PutMode
+    from sqlalchemy import Engine
 
-    from brussels.types.file.file import RemoteMetadataField
+    from brussels.types.file._types import RemoteMetadataField
+
+from brussels.__tests__.types.file.conftest import FakeStoreOps
 
 
 class NoPrimaryKeyMixinModel(Base):
@@ -47,112 +49,6 @@ class TablelessPrimaryKeyModel(PrimaryKeyMixin):
     pass
 
 
-class FakeStoreOps:
-    def __init__(self) -> None:
-        self.calls: list[tuple[str, tuple[object, ...], dict[str, object]]] = []
-        self.put_error: Exception | None = None
-        self.delete_error: Exception | None = None
-        self.put_response: object = {
-            "size_bytes": 5,
-            "content_type": "text/plain",
-            "etag": "etag-123",
-            "checksum": "sum-123",
-            "version": "v1",
-        }
-
-    def _record(self, name: str, args: tuple[object, ...], kwargs: dict[str, object]) -> None:
-        self.calls.append((name, args, kwargs))
-
-    def put(
-        self,
-        path: str,
-        file: object,
-        *,
-        attributes: object | None = None,
-        tags: dict[str, str] | None = None,
-        mode: object | None = None,
-        use_multipart: bool | None = None,
-        chunk_size: int = 5 * 1024 * 1024,
-        max_concurrency: int = 12,
-    ) -> object:
-        self._record(
-            "put",
-            (path, file),
-            {
-                "attributes": attributes,
-                "tags": tags,
-                "mode": mode,
-                "use_multipart": use_multipart,
-                "chunk_size": chunk_size,
-                "max_concurrency": max_concurrency,
-            },
-        )
-        if self.put_error is not None:
-            raise self.put_error
-        return self.put_response
-
-    async def put_async(
-        self,
-        path: str,
-        file: object,
-        *,
-        attributes: object | None = None,
-        tags: dict[str, str] | None = None,
-        mode: object | None = None,
-        use_multipart: bool | None = None,
-        chunk_size: int = 5 * 1024 * 1024,
-        max_concurrency: int = 12,
-    ) -> object:
-        self._record(
-            "put_async",
-            (path, file),
-            {
-                "attributes": attributes,
-                "tags": tags,
-                "mode": mode,
-                "use_multipart": use_multipart,
-                "chunk_size": chunk_size,
-                "max_concurrency": max_concurrency,
-            },
-        )
-        if self.put_error is not None:
-            raise self.put_error
-        return self.put_response
-
-    def get(self, path: str, *, options: object | None = None) -> object:
-        self._record("get", (path,), {"options": options})
-        return b"downloaded-sync"
-
-    async def get_async(self, path: str, *, options: object | None = None) -> object:
-        self._record("get_async", (path,), {"options": options})
-        return b"downloaded-async"
-
-    def get_range(self, path: str, *, start: int, end: int | None = None, length: int | None = None) -> object:
-        self._record("get_range", (path,), {"start": start, "end": end, "length": length})
-        return b"range-sync"
-
-    async def get_range_async(
-        self,
-        path: str,
-        *,
-        start: int,
-        end: int | None = None,
-        length: int | None = None,
-    ) -> object:
-        self._record("get_range_async", (path,), {"start": start, "end": end, "length": length})
-        return b"range-async"
-
-    def delete(self, paths: str | tuple[str, ...] | list[str]) -> None:
-        self._record("delete", (paths,), {})
-        if self.delete_error is not None:
-            raise self.delete_error
-
-    async def delete_async(self, paths: str | tuple[str, ...] | list[str]) -> None:
-        self._record("delete_async", (paths,), {})
-        if self.delete_error is not None:
-            raise self.delete_error
-
-
 def _configure_store(store_ops: FakeStoreOps) -> None:
     remote_storage = cast("RemoteStorage", FileModel.__table__.c["file"].type)
     remote_storage.store = store_ops
@@ -160,13 +56,6 @@ def _configure_store(store_ops: FakeStoreOps) -> None:
 
 def _file_handle(model: FileModel) -> RemoteFile:
     return RemoteFile.from_metadata(model, FileModel.file)
-
-
-@pytest.fixture
-def engine() -> Engine:
-    engine = create_engine("sqlite:///:memory:")
-    DataclassBase.metadata.create_all(engine)
-    return engine
 
 
 def test_from_metadata_rejects_models_without_primary_key_mixin() -> None:
